vehicle-service/requirements.txt
flask~=3.1.1
flask-pymongo
python-dotenv~=1.1.0
flask-cors~=6.0.0

pymongo~=4.13.0
marshmallow~=4.0.0
py-eureka-client~=0.11.0

requests~=2.32.4